refactor: move history to /store
