+++ src/bridge/components/StatelessTransaction.js
@@ -10,7 +10,7 @@ import { Warning } from "../components/Base";
 
 import { BRIDGE_ERROR } from "../lib/error";
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import * as tank from "../lib/tank";
 
 import {
