+++ src/bridge/views/Keystore.js
@@ -8,7 +8,7 @@ import { Row, Col, H1, H3, Warning } from "../components/Base";
 
 import { BRIDGE_ERROR } from "../lib/error";
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import { EthereumWallet } from "../lib/wallet";
 
 class Keystore extends React.Component {
