+++ src/bridge/views/Mnemonic.js
@@ -11,7 +11,7 @@ import {
 import { Row, Col, H1 } from "../components/Base";
 
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import { DEFAULT_HD_PATH, walletFromMnemonic } from "../lib/wallet";
 
 class Mnemonic extends React.Component {
