+++ src/bridge/views/Points.js
@@ -6,7 +6,7 @@ import * as azimuth from "azimuth-js";
 import PointList from "../components/PointList";
 import { NETWORK_NAMES } from "../lib/network";
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import { ETH_ZERO_ADDR, eqAddr } from "../lib/wallet";
 
 const hasTransferProxy = (cache, point) =>
