+++ src/bridge/views/PrivateKey.js
@@ -5,7 +5,7 @@ import { RequiredInput, InnerLabel, InputCaption } from "../components/Base";
 import { Row, Col, H1 } from "../components/Base";
 
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import { EthereumWallet } from "../lib/wallet";
 
 class PrivateKey extends React.Component {
