+++ src/bridge/views/Ticket.js
@@ -13,7 +13,7 @@ import { Row, Col, H1, P } from "../components/Base";
 
 import { randomPatq } from "../lib/lib";
 import { ROUTE_NAMES } from "../lib/routeNames";
-import { withHistory } from "../lib/history";
+import { withHistory } from "../store/history";
 import { urbitWalletFromTicket } from "../lib/wallet";
 
 class Ticket extends React.Component {
